+++ andhow-core/src/main/java/org/yarnandtail/andhow/load/std/StdMainStringArgsLoader.java
@@ -1,6 +1,8 @@
 package org.yarnandtail.andhow.load.std;
 
-import org.yarnandtail.andhow.api.StandardLoader;
+import org.yarnandtail.andhow.api.*;
+import org.yarnandtail.andhow.internal.PropertyConfigurationInternal;
+import org.yarnandtail.andhow.load.BaseKeyValuePairLoader;
 import org.yarnandtail.andhow.load.KeyValuePairLoader;
 
 /**
@@ -62,7 +64,7 @@ import org.yarnandtail.andhow.load.KeyValuePairLoader;
  * {@code AndHowConfiguration.insertLoaderBefore/After()}.
  *
  */
-public class StdMainStringArgsLoader extends KeyValuePairLoader 
+public class StdMainStringArgsLoader extends BaseKeyValuePairLoader
 		implements StandardLoader {
 
 	/**
@@ -71,6 +73,18 @@ public class StdMainStringArgsLoader extends KeyValuePairLoader
 	 */
 	public StdMainStringArgsLoader() {
 		unknownPropertyAProblem = false;
+	}
+
+	@Override
+	public LoaderValues load(final PropertyConfigurationInternal runtimeDef, final ValidatedValuesWithContext existingValues) {
+		return null;	//TODO Remove me
+	}
+
+	@Override
+	public LoaderValues load(final PropertyConfigurationInternal runtimeDef,
+			final LoaderEnvironment environment, final ValidatedValuesWithContext existingValues) {
+
+		return load(runtimeDef, environment.getMainArgs(), KVP_DELIMITER);
 	}
 
 	@Override
